+++ examples/Acl.ts
@@ -1,0 +1,9 @@
+import { Permissions } from '../src/index'
+
+export default class Acl extends Permissions {
+
+  constructor(acl?: any)
+  {
+    super(acl)
+  }
+}