+++ src/plugins/StoreApi.ts
@@ -1,8 +1,7 @@
 import Api from '../api/Api'
 
-// TODO: Testing generic api
 export default class StoreApi extends Api {
-  protected resource = 'v1/eloquent-api/stores'
+  protected resource = 'eloquent-api/stores'
 
   constructor() {
     super()