+++ CHANGELOG.md
@@ -5,11 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.0]
 ### Added
 - Static methods to ApiQuery class
+- Tests
+- Pinia API Store
 ### Changed
+- Removed batch methods from Model
 ### Fixed
+- Validator's $valid response
 
 ## [0.9.2] - 2023-10-08
 ### Added